[WV-2287] Moved modal logic to the modal files [MERGE READY]
