+++ src/js/pages/More/ManageMyCandidates.jsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { Suspense, useCallback, useRef, useState } from 'react';
 import {createPortal} from 'react-dom';
 import styled from 'styled-components';
 import {
@@ -23,48 +23,23 @@ export default function ManageMyCandidates ({ selectedPolitician, selectedPoliti
 
 We’d like to invite you to join WeVote to help support ${selectedPolitician?.politician_name || 'our campaign'}.
 Thanks for your help!`);
-  // Edit modal
+
+  // Modal visibility
   const [showEdit, setShowEdit] = useState(false);
-  const [draftInvite, setDraftInvite] = useState(invitationBody);
-  const [initialInvite, setInitialInvite] = useState(invitationBody);
-  // Preview modal
   const [showPreview, setShowPreview] = useState(false);
-  const handlePreviewOpen = () => setShowPreview(true);
-  const handlePreviewClose = () => setShowPreview(false);
-  // Upload CSV modal
   const [showUpload, setShowUpload] = useState(false);
-  // CSV upload and paste
   const [showPaste, setShowPaste] = useState(false);
-  const [pasteText, setPasteText] = useState('');
-  const [pasteErrors, setPasteErrors] = useState([]);
-  const [mirrorHTML, setMirrorHTML] = useState('');
+
   const [importedVoters, setImportedVoters] = useState([]);
+
   // One-by-one inputs
   const [oneName, setOneName] = useState('');
   const [oneEmail, setOneEmail] = useState('');
   const [onePhone, setOnePhone] = useState('');
-  const fileInputRef = useRef(null);
-  const [allColumnsOK, setAllColumnsOK] = useState(false);
 
-  const openUploadModal = () => { setAllColumnsOK(false); setShowUpload(true); };
-  const closeUploadModal = () => { setAllColumnsOK(false); setShowUpload(false); };
-  const handleSelectCSV = () => fileInputRef.current?.click();
-
-  // Reset draft when invitationBody or showEdit changes
-  useEffect(() => {
-    if (!showEdit) {
-      setDraftInvite(invitationBody);
-      setInitialInvite(invitationBody);
-    }
-  }, [invitationBody, showEdit, setDraftInvite, setInitialInvite]);
+  // Toast notifications
   const [copiedMsg, setCopiedMsg] = useState('');
   const [isSuccessToast, setIsSuccessToast] = useState(false);
-  const openEditModal = () => { setInitialInvite(draftInvite); setShowEdit(true); };
-  const handleEditInvite = () => {
-    setShowPreview(false);
-    openEditModal();
-  };
-
   const toastTimerRef = useRef(null);
   const idRef = useRef(1);
   const makeVoterRecord = useCallback((partial, source = 'Paste list') => ({
@@ -80,71 +55,45 @@ Thanks for your help!`);
     setCopiedMsg(msg);
     toastTimerRef.current = setTimeout(() => { setCopiedMsg(''); setIsSuccessToast(false); }, 2200);
   }, []);
-  const handleInviteSelected = useCallback((rows) => { notify(`Invited ${rows.length} voter${rows.length === 1 ? '' : 's'} by email.`); }, [notify]);
-  const handleInviteEmailOneOrMany = useCallback((rows) => { notify(`Email invite sent to ${rows.length} voter${rows.length === 1 ? '' : 's'}.`); }, [notify]);
-  const handleInviteTextOneOrMany = useCallback((rows) => { notify(`Text invite sent to ${rows.length} voter${rows.length === 1 ? '' : 's'}.`); }, [notify]);
-  const handleHideOne = useCallback((row) => { setImportedVoters((p) => p.filter((r) => (r.id || r._idx) !== (row.id || row._idx))); notify('Hidden from list.', true); }, [notify]);
-  const handleHideMany = useCallback((rows) => { const ids = new Set(rows.map((r) => r.id || r._idx)); setImportedVoters((p) => p.filter((r) => !ids.has(r.id || r._idx))); notify(`Hidden ${rows.length} item${rows.length === 1 ? '' : 's'}.`, true); }, [notify]);
-  const handleSaveInvite = () => {
-    setInvitationBody(draftInvite);
+
+  const handlePreviewOpen = () => setShowPreview(true);
+  const handlePreviewClose = () => setShowPreview(false);
+  const openEditModal = () => setShowEdit(true);
+  const handleEditInvite = () => {
+    setShowPreview(false);
+    openEditModal();
+  };
+
+  const handleSaveInvitation = (newInvitationBody) => {
+    setInvitationBody(newInvitationBody);
     setShowEdit(false);
     notify('Invitation updated.', true);
   };
 
   const emailRE = /\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}\b/i;
 
-  const escapeHTML = (s) => s.replace(/[&<>]/g, (c) => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;' }[c]));
+  const handleInviteSelected = useCallback((rows) => {
+    notify(`Invited ${rows.length} voter${rows.length === 1 ? '' : 's'} by email.`);
+  }, [notify]);
 
-  function parsePastedList (text) {
-    const lines = text.split(/\r?\n/);
-    const rows = [];
-    const errors = [];
+  const handleInviteEmailOneOrMany = useCallback((rows) => {
+    notify(`Email invite sent to ${rows.length} voter${rows.length === 1 ? '' : 's'}.`);
+  }, [notify]);
 
-    lines.forEach((raw, idx) => {
-      const line = raw.trim();
-      if (!line) return;
+  const handleInviteTextOneOrMany = useCallback((rows) => {
+    notify(`Text invite sent to ${rows.length} voter${rows.length === 1 ? '' : 's'}.`);
+  }, [notify]);
 
-      const parts = line.split(',').map((s) => s.trim()).filter(Boolean);
-      const emailCount = (line.match(emailRE) || []).length;
+  const handleHideOne = useCallback((row) => {
+    setImportedVoters((p) => p.filter((r) => (r.id || r._idx) !== (row.id || row._idx)));
+    notify('Hidden from list.', true);
+  }, [notify]);
 
-      if (parts.length > 3) {
-        errors.push({ line: idx, reason: 'Too many commas' });
-        return;
-      }
-      if (emailCount > 1) {
-        errors.push({ line: idx, reason: 'Two emails on one line (missing line break?)' });
-        return;
-      }
-      // Must have at least "Name, Email"
-      if (parts.length < 2) {
-        errors.push({ line: idx, reason: 'Missing email (format: Name, email[, phone])' });
-        return;
-      }
-      // Validate email
-      const emailPart = parts[1].replace(/[<>]/g, '');
-      if (!emailRE.test(emailPart)) {
-        errors.push({ line: idx, reason: 'Invalid email' });
-        return;
-      }
-
-      rows.push({
-        name: parts[0] || '',
-        email: emailPart || '',
-        phone: parts[2] || '',
-      });
-    });
-
-    return { rows, errors };
-  }
-
-  function buildMirrorHTML (text, errors) {
-    const errSet = new Set(errors.map((e) => e.line));
-    return text.split(/\r?\n/).map((line, i) => {
-      const isError = errSet.has(i) && line.trim().length > 0;
-      const safe = escapeHTML(line);
-      return isError ? `<span class="err">${safe}</span>` : safe;
-    }).join('\n');
-  }
+  const handleHideMany = useCallback((rows) => {
+    const ids = new Set(rows.map((r) => r.id || r._idx));
+    setImportedVoters((p) => p.filter((r) => !ids.has(r.id || r._idx)));
+    notify(`Hidden ${rows.length} item${rows.length === 1 ? '' : 's'}.`, true);
+  }, [notify]);
 
   const handleCopyInviteBody = async () => {
     try {
@@ -153,21 +102,6 @@ Thanks for your help!`);
     } catch {
       notify('Copy failed. Select the text and copy manually.', false, 3000);
     }
-  };
-
-  const handleDownloadSample = () => {
-    const csv =
-      'Name,Email,Mobile,Address\n' +
-      'John Smith,js@gmail.com,"(123) 456-7890","123 State St, Anytown, CA 94117"\n';
-    const blob = new Blob([csv], { type: 'text/csv;charset=utf-8;' });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = 'wevote_sample.csv';
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    URL.revokeObjectURL(url);
   };
 
   const handleImportOne = useCallback(() => {
@@ -184,105 +118,19 @@ Thanks for your help!`);
 
     setOneName(''); setOneEmail(''); setOnePhone('');
     notify('Voter added. Ready to invite.', true);
-  }, [oneName, oneEmail, onePhone, notify, makeVoterRecord, setImportedVoters]);
+  }, [oneName, oneEmail, onePhone, notify, makeVoterRecord, emailRE]);
 
-  const handleCSVSelected = async (e) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-
-    if (!/\.csv$/i.test(file.name)) {
-      if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-      setIsSuccessToast(false);
-      setCopiedMsg('Please choose a .csv file.');
-      toastTimerRef.current = setTimeout(() => setCopiedMsg(''), 2500);
-      e.target.value = '';
-      return;
-    }
-    setShowUpload(true);
-
-    const text = await file.text();
-    const lines = text.replace(/\r/g, '').split('\n').filter((l) => l.trim().length > 0);
-    const [headerLine, ...dataLines] = lines;
-    const headers = headerLine.split(',').map((h) => h.trim().toLowerCase());
-
-    const nameIdx = headers.indexOf('name');
-    const emailIdx = headers.indexOf('email');
-    const mobileIdx = headers.indexOf('mobile');
-    const phoneIdx = headers.indexOf('phone');
-    const phoneCol = mobileIdx > -1 ? mobileIdx : phoneIdx;
-    const addressIdx = headers.indexOf('address');
-
-    const ok = nameIdx > -1 && emailIdx > -1 && phoneCol > -1 && addressIdx > -1;
-
-    const rows = dataLines.map((line) => {
-      const cols = line.split(',').map((supporter) => supporter.trim());
-      return {
-        name: nameIdx > -1 ? cols[nameIdx] : '',
-        email: emailIdx > -1 ? cols[emailIdx] : '',
-        phone: phoneCol > -1 ? cols[phoneCol] : '',
-        address: addressIdx > -1 ? cols[addressIdx] : '',
-      };
-    }).filter((r) => r.name || r.email || r.phone || r.address);
-
-    if (ok && rows.length > 0) {
-      setImportedVoters(rows.map((r) => makeVoterRecord(r, 'CSV upload')));
-      closeUploadModal();
-      if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-      setIsSuccessToast(true);
-      setCopiedMsg('All of your columns will be imported.');
-      toastTimerRef.current = setTimeout(() => {
-        setCopiedMsg('');
-        setIsSuccessToast(false);
-      }, 2200);
-    } else {
-      setAllColumnsOK(ok);
-      if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-      setIsSuccessToast(false);
-      if (!ok) {
-        setCopiedMsg('We could not find all required columns (Name, Email, Mobile/Phone, Address). Please adjust and re-upload.');
-      } else if (rows.length === 0) {
-        setCopiedMsg('No rows found in this CSV.');
-      }
-      toastTimerRef.current = setTimeout(() => setCopiedMsg(''), 3000);
-    }
-
-    e.target.value = '';
-  };
-
-  const handlePasteList = () => {
-    setShowPaste(true);
-    setPasteErrors([]);
-    setMirrorHTML('');
-  };
-
-  const closePaste = () => {
-    setShowPaste(false);
-    setPasteErrors([]);
-    setMirrorHTML('');
-  };
-
-  const onPasteTextChange = (e) => {
-    const { value } = e.target;
-    setPasteText(value);
-    const { errors } = parsePastedList(value);
-    setPasteErrors(errors);
-    setMirrorHTML(buildMirrorHTML(value, errors));
-  };
-
-  const handlePasteImport = () => {
-    const { rows, errors } = parsePastedList(pasteText);
-    if (errors.length) {
-      setPasteErrors(errors);
-      setMirrorHTML(buildMirrorHTML(pasteText, errors));
-      return; // block import until fixed
-    }
+  const handleImportFromPaste = (rows) => {
     setImportedVoters((prev) => [...prev, ...rows.map((r) => makeVoterRecord(r, 'Paste list'))]);
-    if (toastTimerRef.current) clearTimeout(toastTimerRef.current);
-    setCopiedMsg(`Imported ${rows.length} voter${rows.length !== 1 ? 's' : ''} from pasted list.`);
-    toastTimerRef.current = setTimeout(() => setCopiedMsg(''), 2200);
+    notify(`Imported ${rows.length} voter${rows.length !== 1 ? 's' : ''} from pasted list.`, true);
+    setShowPaste(false);
   };
 
-  const prospectiveCount = useMemo(() => parsePastedList(pasteText).rows.length, [pasteText]);
+  const handleImportFromCSV = (rows) => {
+    setImportedVoters(rows.map((r) => makeVoterRecord(r, 'CSV upload')));
+    setShowUpload(false);
+    notify('All of your columns will be imported.', true);
+  };
 
   return (
     <>
@@ -330,12 +178,12 @@ Thanks for your help!`);
       <Section>
         <H3>Upload voters from a file or paste a list</H3>
         <Row>
-          <PillButton type="button" onClick={openUploadModal}>
+          <PillButton type="button" onClick={() => setShowUpload(true)}>
             <UploadIcon fontSize="small" />
             Upload CSV file
           </PillButton>
           <Or>OR</Or>
-          <PillButton type="button" onClick={handlePasteList}>
+          <PillButton type="button" onClick={() => setShowPaste(true)}>
             <PasteListIcon size={22} />
             Paste list
           </PillButton>
@@ -372,38 +220,23 @@ Thanks for your help!`);
       <EditInvitationModal
         isOpen={showEdit}
         onClose={() => setShowEdit(false)}
-        draftInvite={draftInvite}
-        setDraftInvite={setDraftInvite}
-        initialInvite={initialInvite}
-        onSave={handleSaveInvite}
+        invitationBody={invitationBody}
+        onSave={handleSaveInvitation}
         notify={notify}
         selectedPoliticianId={selectedPoliticianWeVoteId}
       />
       {/* Paste list modal */}
       <PasteListModal
         isOpen={showPaste}
-        onClose={closePaste}
-        pasteText={pasteText}
-        onPasteTextChange={onPasteTextChange}
-        mirrorHTML={mirrorHTML}
-        pasteErrors={pasteErrors}
-        onImport={handlePasteImport}
-        prospectiveCount={prospectiveCount}
+        onClose={() => setShowPaste(false)}
+        onImport={handleImportFromPaste}
       />
       {/* Upload CSV modal */}
       <UploadCSVModal
         isOpen={showUpload}
-        onClose={closeUploadModal}
-        onDownloadSample={handleDownloadSample}
-        onSelectFile={handleSelectCSV}
-        allColumnsOK={allColumnsOK}
-      />
-      <input
-        ref={fileInputRef}
-        type="file"
-        accept=".csv"
-        style={{ display: 'none' }}
-        onChange={handleCSVSelected}
+        onClose={() => setShowUpload(false)}
+        onImport={handleImportFromCSV}
+        notify={notify}
       />
       {copiedMsg && createPortal(
         <Toast role="status" aria-live="polite" $success={isSuccessToast}>
